Migrating `MetricDetails` to typed functional component.
